refactor: refatorado e adicionado novos métodos nos repositórios in memory.
