+++ src/clientes/infrastructure/persistence/clienteInMemory.database.ts
@@ -1,7 +1,8 @@
 import { Cliente } from '../../domain/cliente.entity';
+import { randomUUID } from 'crypto';
 export class InMemoryClientesDatabase {
   public database: Cliente[] = [
-    new Cliente('002224e-0251jk', 'João', '12345678900', 'joao@ex.com', '11947474747', '08710000', '123456'),
-    new Cliente('0144ee-kso2514', 'Maria', '98765432100', 'maria@ex.com', '11936363636', '08710000', '123456'),
+    new Cliente(randomUUID(), 'joao', '01234567890', 'joao@ex.com', 'joao1234', '11474548520', '08710000', '123456'),
+    new Cliente(randomUUID(), 'maria', '98765432100', 'maria@ex.com', 'maria123', '11235654895', '08710000', '123456'),
   ];
 }
