+++ src/gerentes/infrastructure/persistence/gerenteInMemory.database.ts
@@ -1,5 +1,7 @@
 import { Gerente } from '../../domain/gerente.entity';
 
 export class InMemoryGerenteDatabase {
-  public database: Gerente[] = [new Gerente('a5s5a15-asa551155', '12214714', 'Ana', '11958457754', 'ana@bancocaqui.com', '11958584752')];
+  public database: Gerente[] = [
+    new Gerente('dlllfd55-dsdssd', '12345678', 'nono', '30000000144', 'ana@bancocaqui.com', 'ana12345', '11999999999'),
+  ];
 }
